pamsm: add an optional libpam feature
This adds an optional `libpam` feature to enable linking against and
using native C methods from `libpam.so`. As this provides additional
public API and cargo features are additive, this is opt-in and not
on by default.
This also means that consumers of this crate will not get anymore a
link dependency to `libpam.so`, unless they enable this feature.