pam_types: align ffi types with upstream
PamMsgStyle is not in C format and breaks when talking to libpam, this
has been corrected by tagging it as #[repr(C)]. Similarly,
PamMessage.msg is listed as *const u8, despite upstream calling it a
char*. This works because they are the same length, but makes rustc have
a hard time. Setting this to what upstream expects, *const c_char, works
as expected.