pam: implement conv
Currently there is no way to do custom communications via PamLibExt.
This change exposes pam_conv's callback as a custom function and
pam_types::PamMsgStyle, since it is a required parameter.

A common CString failure message was broken out as a constant, and seems
to trigger clippy's dead_code lint. The PamResponse impl has been
removed, since it was unused and broken during type modifications. This
also allowed the removal of the libc crate dependency.